requirements.txt
python-telegram-bot
joblib
langdetect
python-dotenv